requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_bcrypt
psycopg2-binary
shortuuid
Pillow